chat-service/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
pika==1.3.2
google-genai==1.23.0
supabase==2.16.0
python-dotenv==1.1.1
asyncio==3.4.3
aio-pika==9.5.5
numpy==2.3.1
